generic object implementation

+++ include/utils.h
@@ -5,8 +5,8 @@
 # include <stdbool.h>
 # include <stdlib.h>
 # include <stdio.h>
-#include <fcntl.h>
-#include <sys/types.h>
+# include <fcntl.h>
+# include <sys/types.h>
 
 # include "mlx.h"
 # include "mlx_int.h"
@@ -37,14 +37,14 @@ void		**ft_newarr(void);
 
 // Colors
 int			set_rgba(t_vec4 c);
-t_vec4		get_rgba(float color);
+t_vec4		get_rgba(int color);
 int			valid_color(t_vec4 c);
 
 // mlx wrappers
 typedef struct s_canvas	t_canvas;
 
-void	*ft_getmlx(void);
-void	*ft_new_image(int w, int h);
-void	*ft_new_windows(int w, int h, char *name);
+void		*ft_getmlx(void);
+void		*ft_new_image(int w, int h);
+void		*ft_new_windows(int w, int h, char *name);
 
 #endif // UTILS_H
